fix(): puntos extra son decimales (0.00-5.00) y se suman literal a la calif /10
- courses.extra_points pasa de Integer a Numeric(3,2). la migracion
  c3d4e5f6a7b8 se edito en sitio (aun no aplicada en prod)
- CourseUpdate/CourseOut: extra_points ahora float, ge=0 le=5
- analisis: withExtra = base10 + extra_points (tal cual, capado a 10),
  ya no /10. input con step=0.5. label "Puntos extra"

+++ backend/src/courses/schemas.py
@@ -34,14 +34,16 @@ class CourseUpdate(BaseModel):
     # con los endpoints /categories.
     name: str | None = None
     # Ajustes de la pestaña Análisis:
-    extra_points: int | None = Field(default=None, ge=0, le=5)
+    # puntos extra: decimal 0.00 - 5.00 (p. ej. 3.5), se suman tal cual a la
+    # calificación en escala 0-10.
+    extra_points: float | None = Field(default=None, ge=0, le=5)
     rounding_enabled: bool | None = None
     rounding_method: RoundingMethod | None = None
 
 
 class CourseOut(CourseBase):
     id: int
-    extra_points: int = 0
+    extra_points: float = 0
     rounding_enabled: bool = True
     rounding_method: RoundingMethod = "half_up"
     model_config = ConfigDict(from_attributes=True)
